Separate the Logo from the Title with a Title component
